requirements.txt
discord.py==1.2.1
tabulate==0.8.3
peewee==3.9.5